Subdomain fix
BREAKING CHANGE: The way SANs and the main domain are handled has been completely revamped. This breaks the old way of providing subdomains (as a map of lists), and the way in which the main domain was set. None of this worked correctly before, and the ways in which subdomains and the main domain is handled fixes everything.

- New var.main_subdomain added
- var.subdomains is now a list
- var.parent_zone will now be used as the "main subdomain" when neither var.subdomains or var.main_subdomain are provided
- Only a single certificate will be generated per instance of the module

See the examples for updated usage.